Move AMA pages into AMAs folder
